refactor(project): extract storage/backups/delivery/uploads concerns
Splits Project's storage, backup, delivery and upload behavior into focused
concerns and dedupes the connection-prefixed key building into one storage_key
helper shared by backups and delivery.
